Move weighted-pool v3 to /deprecated
