+++ pkg/deployments/tasks/deprecated/20230206-weighted-pool-v3/input.ts
@@ -1,4 +1,4 @@
-import Task, { TaskMode } from '../../src/task';
+import Task, { TaskMode } from '../../../src/task';
 
 export type WeightedPoolDeployment = {
   Vault: string;
