+++ pkg/deployments/tasks/deprecated/20230206-weighted-pool-v3/output/bsc.json
