+++ pkg/deployments/tasks/deprecated/20230206-weighted-pool-v3/readme.md
@@ -1,6 +1,9 @@
 # 2023-02-06 - Weighted Pool Factory V3
 
-Deployment of the `WeightedPoolFactory`, which allows creating Weighted Pools.
+> ⚠️ **DEPRECATED: do not use** ⚠️
+> This deployment has been deprecated in favor of a new version that uses create2 for pool deployments: [weighted-pool-v4](../../20230320-weighted-pool-v4/).
+
+Deployment of the `WeightedPoolFactory`, which allows creating Weighted Pools (with create2).
 Supersedes `20220908-weighted-pool-v2`, fixing the reentrancy issue described in this [forum post](https://forum.balancer.fi/t/reentrancy-vulnerability-scope-expanded/4345).
 
 ## Useful Files
