+++ pkg/deployments/tasks/deprecated/20230206-weighted-pool-v3/test/contracts/ReadOnlyReentrancyAttackerWP.sol
